move erc20 contract
