+++ app/page/home/activity.tsx
@@ -5,10 +5,10 @@ import ArrowUpRightFromSquare from "react:~assets/arrowUpRightFromSquare.svg"
 import CircleXmark from "react:~assets/circleXmark.svg"
 import Clock from "react:~assets/clock.svg"
 
-import { ERC20Contract } from "~/packages/contract/erc20"
 import { ScrollableWrapper } from "~app/component/scrollableWrapper"
 import { useAccount, useNetwork, useTransactionLogs } from "~app/hook/storage"
 import { decodeExecuteParams } from "~app/util/calldata"
+import { ERC20Contract } from "~packages/eip/20/contract"
 import { getChainName } from "~packages/network/util"
 import address from "~packages/util/address"
 import number from "~packages/util/number"
